pause and resume consumption

+++ thoth/investigator/configuration.py
@@ -46,7 +46,7 @@ class Configuration:
     PENDING_WORKFLOW_LIMIT = os.getenv("ARGO_PENDING_WORKFLOW_LIMIT", None)
 
     # Consumer configuration
-    MAX_RETRIES = float(os.getenv("THOTH_INVESTIGATOR_MAX_RETRIES", 5))
+    MAX_RETRIES = int(os.getenv("THOTH_INVESTIGATOR_MAX_RETRIES", 5))
     BACKOFF = float(os.getenv("THOTH_INVESTIGATOR_BACKOFF", 0.5))  # Linear backoff strategy
     ACK_ON_FAIL = bool(int(os.getenv("THOTH_INVESTIGATOR_ACK_ON_FAIL", 0)))
     NUM_WORKERS = int(os.getenv("THOTH_CONSUMER_WORKERS", 5))
